SONARJAVA-6473 Extract CFG block helpers
