+++ java-checks/src/main/java/org/sonar/java/checks/RedundantJumpCheck.java
@@ -17,25 +17,19 @@
 package org.sonar.java.checks;
 
 import java.util.Arrays;
-import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
-import java.util.Set;
 import org.sonar.check.Rule;
 import org.sonar.java.cfg.CFG;
 import org.sonar.java.cfg.CFG.Block;
-import org.sonar.java.model.ExpressionUtils;
+import org.sonar.java.cfg.CFGUtils;
 import org.sonar.plugins.java.api.IssuableSubscriptionVisitor;
 import org.sonar.plugins.java.api.tree.BlockTree;
-import org.sonar.plugins.java.api.tree.DoWhileStatementTree;
-import org.sonar.plugins.java.api.tree.ExpressionTree;
-import org.sonar.plugins.java.api.tree.ForStatementTree;
 import org.sonar.plugins.java.api.tree.MethodTree;
 import org.sonar.plugins.java.api.tree.ReturnStatementTree;
 import org.sonar.plugins.java.api.tree.StatementTree;
 import org.sonar.plugins.java.api.tree.Tree;
 import org.sonar.plugins.java.api.tree.TryStatementTree;
-import org.sonar.plugins.java.api.tree.WhileStatementTree;
 
 @Rule(key = "S3626")
 public class RedundantJumpCheck extends IssuableSubscriptionVisitor {
@@ -67,10 +61,10 @@ public class RedundantJumpCheck extends IssuableSubscriptionVisitor {
       && !isSwitchCaseChild(terminator)
       && hasValidContinueTarget(terminator)) {
 
-      successorWithoutJump = nonEmptySuccessor(successorWithoutJump);
+      successorWithoutJump = CFGUtils.nonEmptySuccessor(successorWithoutJump);
       Iterator<Block> successors = block.successors().iterator();
       if (successors.hasNext()) {
-        Block successor = nonEmptySuccessor(successors.next());
+        Block successor = CFGUtils.nonEmptySuccessor(successors.next());
         if (successor.equals(successorWithoutJump)
           && !isJumpThroughFinallyWithDistinctContinuation(terminator, successor)) {
           reportIssue(terminator, "Remove this redundant jump.");
@@ -109,33 +103,8 @@ public class RedundantJumpCheck extends IssuableSubscriptionVisitor {
     return tree.is(Tree.Kind.WHILE_STATEMENT, Tree.Kind.DO_STATEMENT, Tree.Kind.FOR_STATEMENT, Tree.Kind.FOR_EACH_STATEMENT);
   }
 
-  private static Block nonEmptySuccessor(Block initialBlock) {
-    Block result = initialBlock;
-    Set<Integer> visited = new HashSet<>();
-    while (isEffectivelyEmpty(result) && result.successors().size() == 1 && visited.add(result.id())) {
-      result = result.successors().iterator().next();
-    }
-    return result;
-  }
-
-  private static boolean isEffectivelyEmpty(Block block) {
-    return block.elements().stream().allMatch(element -> element.is(Tree.Kind.EMPTY_STATEMENT));
-  }
-
-  private static boolean isFinallyBlockWithDistinctContinuation(Block block) {
-    Block exitBlock = block.exitBlock();
-    if (!block.isFinallyBlock() || exitBlock == null) {
-      return false;
-    }
-    Block successorAfterJump = nonEmptySuccessor(exitBlock);
-    return block.successors().stream()
-      .filter(successor -> !isDeadLoopExitingTo(successor, exitBlock))
-      .map(RedundantJumpCheck::nonEmptySuccessor)
-      .anyMatch(successor -> !successor.equals(successorAfterJump));
-  }
-
   private static boolean isJumpThroughFinallyWithDistinctContinuation(Tree terminator, Block successor) {
-    if (!isFinallyBlockWithDistinctContinuation(successor)) {
+    if (!CFGUtils.isFinallyBlockWithDistinctContinuation(successor)) {
       return false;
     }
     if (terminator.is(Tree.Kind.RETURN_STATEMENT)) {
@@ -180,19 +149,4 @@ public class RedundantJumpCheck extends IssuableSubscriptionVisitor {
         .anyMatch(s -> !s.is(Tree.Kind.EMPTY_STATEMENT));
   }
 
-  private static boolean isDeadLoopExitingTo(Block successor, Block exitBlock) {
-    Tree terminator = successor.terminator();
-    if (terminator == null || !exitBlock.equals(successor.falseBlock())) {
-      return false;
-    }
-    ExpressionTree condition = null;
-    if (terminator.is(Tree.Kind.DO_STATEMENT)) {
-      condition = ((DoWhileStatementTree) terminator).condition();
-    } else if (terminator.is(Tree.Kind.FOR_STATEMENT)) {
-      condition = ((ForStatementTree) terminator).condition();
-    } else if (terminator.is(Tree.Kind.WHILE_STATEMENT)) {
-      condition = ((WhileStatementTree) terminator).condition();
-    }
-    return condition != null && Boolean.FALSE.equals(ExpressionUtils.resolveAsConstant(condition));
-  }
 }
